perf(优化):1.事件总线添加回调 2.网络消息编解码器修改消息包大小所占位数（8bit->32bit）

+++ gameserver/src/main/java/com/grtwwh/gameserver/base/socket/coder/SocketPacketEncoder.java
@@ -21,7 +21,7 @@ public class SocketPacketEncoder extends MessageToByteEncoder {
         }
         SocketPacket socketPacket = (SocketPacket) msg;
         out.writeInt(socketPacket.getPacketId());
-        out.writeByte(socketPacket.getData().length);
+        out.writeInt(socketPacket.getData().length);
         out.writeBytes(socketPacket.getData());
     }
